web-apps/requirements.txt
beautifulsoup4==4.11.2
blinker==1.6.2
click==8.1.7
Django==4.1.2
email-validator==1.3.1
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.0.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy==1.0.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.31
numpy==1.23.4
openai==0.27.5
pandas==1.5.3
playsound==1.3.0
python-dotenv==1.0.0
redis==4.5.1
requests==2.28.2
requirements-detector==1.0.3
selenium==4.8.2
spotipy==2.22.1
SQLAlchemy==1.4.46
sqlparse==0.4.3
twilio==7.16.3
urllib3==1.26.12
virtualenv==20.23.0
Werkzeug==2.2.3
WTForms==3.0.1
